refactor: remove set_module
