+++ miroslava/__init__.py
@@ -21,7 +21,9 @@ from miroslava.globals import current_app as current_app
 from miroslava.globals import g as g
 from miroslava.globals import request as request
 from miroslava.globals import session as session
+from miroslava.utils import abort as abort
 from miroslava.utils import jsonify as jsonify
+from miroslava.utils import make_response as make_response
 from miroslava.utils import render_template as render_template
 from miroslava.wrappers import Request as Request
 from miroslava.wrappers import Response as Response
